LayoutTests/http/tests/websocket/tests/hybi/inspector/handshake-error-expected.txt

Tests WebSocket handshake errors.


== Running test suite: WebSocket.HandshakeError
-- Running test case: HandshakeError
PASS: Resource should be a WebSocket.
PASS: Should have correct status code.
PASS: Should have correct status text.
